Add FFT module with fft, ifft, and applyTransferFunction methods, remove methods from featuretest
